% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calucate_variable_importance.R
\name{calculate_variable_importance}
\alias{calculate_variable_importance}
\title{Calculate Variable Importance}
\usage{
calculate_variable_importance(models)
}
\arguments{
\item{models}{List of trained models from train_models().}
}
\value{
A list of data frames containing variable importance for each model.
}
\description{
This function calculates and returns the variable importance for each model.
}
