% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalutate_models.R
\name{evaluate_models}
\alias{evaluate_models}
\title{Evaluate Model Performance}
\usage{
evaluate_models(models)
}
\arguments{
\item{models}{List of trained models from train_models().}
}
\value{
A list containing performance metrics and plots.
}
\description{
This function evaluates the performance of trained models using various metrics.
}
