% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_roc_curves}
\alias{plot_roc_curves}
\title{Plot ROC Curves}
\usage{
plot_roc_curves(eval_results)
}
\arguments{
\item{eval_results}{Evaluation results from evaluate_models().}
}
\value{
A ggplot object containing the ROC curves.
}
\description{
This function plots ROC curves for the trained models.
}
