% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_variable_importance}
\alias{plot_variable_importance}
\title{Plot Variable Importance}
\usage{
plot_variable_importance(var_imp, model_name)
}
\arguments{
\item{var_imp}{Variable importance results from calculate_variable_importance().}

\item{model_name}{Name of the model to plot.}
}
\value{
A ggplot object containing the variable importance plot.
}
\description{
This function plots the variable importance for a specified model.
}
