% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for Alzheimer's Biomarker Analysis}
\usage{
prepare_data(file_path)
}
\arguments{
\item{file_path}{Character string specifying the path to the CSV file.}
}
\value{
A list containing the prepared data and other relevant information.
}
\description{
This function loads and prepares the data for analysis, separating it into
Alzheimer's Disease (AD) and Healthy Control (HC) groups.
}
