% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split Data into Training and Test Sets}
\usage{
split_data(prepared_data, train_ratio = 0.7)
}
\arguments{
\item{prepared_data}{List containing the prepared data from prepare_data().}

\item{train_ratio}{Numeric value between 0 and 1 specifying the ratio of
training data.}
}
\value{
A list containing the training and test datasets.
}
\description{
This function splits the prepared data into training and test sets.
}
