% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{train_models}
\alias{train_models}
\title{Train Multiple Machine Learning Models}
\usage{
train_models(train_data, n_cv = 10, n_repeat = 10)
}
\arguments{
\item{train_data}{Training data frame.}

\item{n_cv}{Number of cross-validation folds.}

\item{n_repeat}{Number of repetitions for cross-validation.}
}
\value{
A list of trained models.
}
\description{
This function trains multiple machine learning models on the provided data.
}
