% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlzheiMarkers-package.R
\docType{package}
\name{AlzheiMarkers-package}
\alias{AlzheiMarkers}
\alias{AlzheiMarkers-package}
\title{AlzheiMarkers: Analyze Metabolomic Biomarkers for Alzheimer's Disease}
\description{
This package provides tools for analyzing metabolomic biomarkers to distinguish Alzheimer's Disease (AD) patients from Healthy Controls (HC). It includes functions for data preparation, model training, evaluation, and visualization of results.
}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0009-0001-9514-751X}{ORCID})

}
\keyword{internal}
